pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "public-renewable-grid-resilience-auditor"
version = "0.1.0"
description = "Public-sector renewable integration and grid resilience analytics platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.1.3,<2.0.0",
  "pyarrow>=18.0.0,<20.0.0",
  "pandas>=2.2.0,<3.0.0",
  "requests>=2.32.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pydantic>=2.9.0,<3.0.0",
  "streamlit>=1.40.0,<2.0.0",
  "dbt-duckdb>=1.9.0,<2.0.0"
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
